fix: Twitter API 404 when auth token not set
The previous official API is not usable when not logged in,
now we have two options, one is the possibly legacy API in
Fritter with a different endpoint and authorization token,
the other is the latest official public API for guests
(`/TweetResultByRestId`).
We chose the official public API since it seems to have the same ability
as the Fritter one, while providing clearer error messages.

Now, logged in users will still be using the previous API, while others
will use the public API.

Furthermore, in order to use the same feature flags for different APIs,
we merged the flags and found a sweet point.

close #81